Add unfinished test
